+++ src/index.ts
@@ -1,1 +1,2 @@
 export { getElementSourceLocation } from './getElementSourceLocation'
+export { resolveSourceLocationInServer } from './resolveSourceLocationInServer'